Better handling of ActiveDocument subclasses.
Let's pretend we have a super-simple ActiveDocument arrangement:

  class A
    include ActiveDocument
  end

  class B < A
  end

Calling A.all will retrieve all instances of A and B. However, calling
B.all will only retrieve documents which are an instance of B:

  A.all
  # => [#<A ...>, #<A ...>, #<B ...>]

  B.all
  # => [#<B ...>]

This is consistent with ActiveRecord. It doesn't make a great deal of
sense, however, until you consider that we use it mostly in Nodes:

  - Node
    - StatNode
    - ConverterNode
    - DemandNode
      - FinalDemandNode

... so we can do FinalDemandNode.all and get just those nodes which
implement final demand, etc...

Also, calling "all" on a subclass would incorrectly instantiate
top-level instances (e.g. nodes which were *only* Nodes and not some
other subclass) as instances of the subclass instead of the top-level
class.