Use ETSource.data_dir in ActiveModel.
Rather than referring to ETSource.root and having to stub the
directories for each ActiveDocument class in the tests, the classes will
now use ETSource.data_dir to specify the root of the data/ directory.

Therefore in tests we can wrap specs in ETSource.with_data_dir to
temporarily alter the path to point to the fixture directory.

The use of "with_data_dir" is not yet implemented in the tests, but will
come in a separate commit.