Add ETSource::Runner.
Manages the process of reading the source files (through
ActiveDocument.all), co-ordinating creation of the graph using
GraphBuilder, and calculates Rubel-based attributes.

This is currently limited to node demands.

Runtime has been changed to require that a Graph is provided which will
be required for queries which need graph objects.