Add ActiveDocument "ns" and "ns=".
The document "namespace" is determined by looking at the subdirectory in
which it resides. For example, a document at ./a/b/c/doc.ad will have
the namespace "a.b.c".

You can change the namespace with "ns=".

This will be aliased on some document classes, such as Node#sector.