Changed has_energy_storage into has_electricity_storage
In order to be consistent, this dependable attribute has been renamed to
has_electricity_storage. This change is associated with similar changes
(and commits) to respectively etmodel and etengine.

CAUTION: Following the update of these three repositories, the dependent on
attribute of the associated sidebar and chart needs to be updated as well!!