Add Lignite and Uranium extraction to preset_demand group
Each of these nodes received a preset_demand in the *future* graph as a
result of the "initial_updates_preset_demand" input, but no preset
demand would be set on the *present* graph.

This poses a problem with the new German dataset which has a non-zero
value for Uranium extraction. By adding these two nodes to the
preset_demand group (the other three extraction nodes are already in the
group), ETEngine will set the preset_demand using the value from
Refinery.