Remove "final_demand" type from non-FD nodes
Although many of these nodes contain "final_demand" in their name, and
are in fact final demand nodes in ETEngine, in Atlas+Refinery, the type
is used to indicate that the node document MUST have a query and
associated value in a CSV file. Many of these nodes do not have -- and
are not expected to have -- such a value.

This fixes the failing (actually, skipped) validations on these nodes.
Ref #283.