Execute merit order input before FCE updates
For the UPDATE_FCE inputs to be executed, ETEngine must initialize its
graph plugins; one of which is the merit order plugin. With the FCE and
merit inputs having the same priority, the FCE inputs would be executed
prior to that disabling the merit order. This caused the full merit
order plugin to be turned on, even though the user had disabled it.

Increasing the priority of `settings_enable_merit_order` ensures that it
is executed before the FCE updates, correctly disabling the merit order.

Ref quintel/etmodel#2116