Set custom number of cars in derived datasets
Note that the area attribute is currently not sufficient to do this.
While the intention is to deprecate the area attribute in favour of a
CSV, the custom number of cars must be set with an initializer input
otherwise a scaled value from the original dataset will be used instead.