Fix ETMoses buffer queries
Hot water and space heating nodes are not directly connected to useful
demand nodes, and the use of Array#sum (on nil) would raise an error due
to no matching nodes.

Replaces CHILDREN(useful_demand) by referencing the relevant group names
instead and removes the use of Array#sum in favour of SUM().

Closes quintel/etmoses#1610