Removed edge between refinery production node and heat input
This edge lead to a double-counting in primary demand and CO2 emissions. Energetic use in the Refinery sector (used to make heat) is included in industry final demand and hence should not also be included in the primary demand of e.g. transport fuels used in the transport sector. Removing this edge ensures that only non-energetic oil is taken into account when calculating the primary demand of transport fuels