Updated slider max for electricity and heat plants
The slider max for electricity and heat plants now takes into account installed capacity of the plants in the starting year. In this way the slider max can never be too restrictive to display installed capacity in the present. For electricity plants the slider max is now equal to (the installed capacity required to supply) total electricity demand in the present, or twice the installed capacity of the plant in the present (whichever is larger). For heat plants the slider max equals half the (installed capacity required to supply) total heat production in the present (both collective heat and individual heat), or twice the installed capacity of the plant in the present  (whichever is bigger). Closes https://github.com/quintel/etsource/issues/2085