Dashboard queries should only run in the future
Changes all but one dashboard query so that we short-circuit calculating
a value when run against the present graph. The dashboard only shows the
future value, therefore calculating one for the present wastes time.

The only exception is for the calculation of final demand plus losses,
where the dashboard requires the present value to calculate the delta.