Added geothermal to industry heat network
The geothermal well does not have electricity input. Adding this results in a topology sort error. It looks like the reason for this is that the steam hot water from geothermal can be used in the Refineries sector, which produces diesel, which can be used to generate power, resulting in a loop (geothermal is indirectly involved in E production and has an E demand)