Fix references to old "_and_p2h" useful demand node
P2H was removed in 950a2916c and 5c0c98703 with the
"solar_heater_and_p2h" useful demand node being replaced with just
"solar_heater".

This commit fixes that a number of inputs and derived dataset values
needed to be adjusted to reference the replacement node when querying a
link.