Add waste carriers to CHPs
Added steam hot water as a waste carrier to selected CHPs. This ensures that the fuel and CO2 costs for producing this heat are not included in the marginal (electricity) cost calculation, moving these CHPs up in the Merit Order. This attribute has been added to CHPs with a steam cycle: all CHPs without a gaseous input carrier and all combined cycle CHPs.

See: https://github.com/quintel/etengine/pull/1086