Adjust buildings solar production query
The chart "Sources of electricity in buildings" did not show all solar PV production in the buildings sector as part of it was assumed to go to the ICT sector. This is confusing as the solar PV slider is in the buildings section and the chart suggests to show all PV production. Adjusted the query, it now shows full PV production in the chart