Set industry coal output loss to zero for all local datasets
Set output loss of industry_transformation_generic_coal to zero for all local datasets. This loss percentage is based on the parent dataset to match final and primary demand in the metal sector. As most local datasets do not have a significant metal sector, this loss distorts the primary demand calculation as any coal use in industry leads to a much larger primary demand of coal.