Change imported heat CO2 input statement
The previous statement updated the fossil_heat and renewable_heat properties of the imported_heat.yml FCE file. Since the names of these properties are not necessarily the same across datasets, the CO2 of datasets with differently named properties would not update.

This commit changes the input statement to a similar approach as the imported electricity CO2 input based on node/edge names, making it more robust than the previous approach.

Closes quintel/etmodel#3457