Rename node curve names to ensure customisable curves only affect the intended nodes
Users will be able to overwrite curves using the front-end. This means that all nodes in a customisable category should use the same, unique curve name. Nodes that are not in the same customisable category should not use that curve name.

For example: users will be able to upload a custom curve for 'Charging at home' of electric cars. The corresponding curve name is electric_vehicle_profile_1. Uploading a custom electric_vehicle_profile_1 should only affect the curve of charging-at-home cars. This means that electric_vehicle_profile_1 should not be used by e.g. electric buses or trucks. Therefore, the curve names used in these nodes have been changend to 'electric_buses' and 'electric_trucks' respectively.