Add costs for higher insulation levels
Previously, costs for insulation levels higher than the NL slider max were either zero or empty. This lead to query errors with datasets with a higher slider max that had the NL files symlinked. This commits sets the insulation costs for the missing higher values equal to the highest known value.