Distinguish between refinery products and primary CO2
In line with main CO2 chart, make a distinction between emissions from refinery products (gasoline, diesel etc.) and
other carriers used in transport (electricity, hydrogen, LNG etc.). For the former, we read the emission factor directly from the carrier instead of using the primary_co2_emission method. The reason is that the latter leads to double-counting of the final demand of the refineries sector. Also, it requires very sophisticated refinery modelling to determine the exact amount of primary oil used to produce gasoline, diesel etc.