Change link type between unused heat and dump
See #1173. This link being inversed_flexible opts it in to special
behaviour in Refinery, whereby when demand is zero the link demand (and
therefore steam_hot_water slot share) are defaulted to zero. Instead, we
want the opposite, and will use explicit shared in the "unused heat"
node to accomplish this.