Remove output of not_defined
These nodes were the only 'dead ends' in the graph with a positive output conversion (except 'loss'), i.e. output.not_defined = 1. This resulted a total node output (output_of_not_defined and output_of_loss) that was twice as high as the total input (when querying these inputs and outputs per carrier separately, as is the case in the energy_flows data export). Removing the output of not_defined solves this.