Adjustment to SGQs to reflect metal graph changes
The graph now has separate final demand nodes for steel, aluminium and other metals. This greatly reduces the number and complexity of the SGQs needed to set the graph_values. Based on the user inputs we now only need to determine the total amount of electricity/network gas/coal/steam hot water going to steel and aluminium.