Russian Gas queries don't need to return deltas
The Russian Gas Tool keeps track of the original future value and
compares it with the new value resulting from the user changes. The
present value is never needed.

To that end, the queries have been changed to return amounts in energy
(which the tool will compare) rather than the delta between the future
and present.