Add a config file specifying the default dataset
This configures ETEngine to use the named dataset as the default. The
default is used in the admin pages, and by ETEngine to initialize it's
graph structure and data. Changing this to nl2019 allows users who lack
the ETSource decryption key to still run ETEngine for themself and load
the public datasets.