Move industry final demand edge group to include DSR demand
Before, only one of the edges from buildings local production to the ICT sector was marked with an edge group. The edge to DSR was empty. To solve this, the edge group is moved 'deeper' into the graph, where the DSR edge recombines with the regular demand edge. This way, all ICT demand is include in the final demand edge group.