Apply manual calculation of electric vehicle OPEX and revenue
The node setup and relevant attributes of the deployable capacity of electric vehicles for storage is spread over the regular electric vehicle node and the p2p node. Therefore, applying the cost methods for OPEX and revenue will probably cause issues here. A pragmatic approach is taken here by manually calculating the OPEX and revenue based on electricity in- and output of the p2p nodes and the electricity price curve