Economic performance of electricity storage technologies (#3268)
* Addition of queries for electricity storage table

* Implementation of wtp exception in economic performance storage tabel

* Changes in WTP queries

* Change of cost methods in queries

* Altering revenue method

* Change merit order query unit to number

* Revert "Change merit order query unit to number"

This reverts commit b3c01d566fc2c1230eff36f6274bbb467f912491.

* Update capacity and capex queries

* Apply manual calculation of electric vehicle OPEX and revenue

The node setup and relevant attributes of the deployable capacity of electric vehicles for storage is spread over the regular electric vehicle node and the p2p node. Therefore, applying the cost methods for OPEX and revenue will probably cause issues here. A pragmatic approach is taken here by manually calculating the OPEX and revenue based on electricity in- and output of the p2p nodes and the electricity price curve

* Replacing zero's with dashes in ev revenue and viability queries

* Altering wtp queries

---------

Co-authored-by: Kyra de Haan <kyra.dehaan@quintel.com>