Remove biofootprint group from import of biogenic waste
The typical_production_per_km2 of 0.0 for the biogenic waste carrier causes the biofootprint calculation to break. Previously, due to a typo this node was excluded from the biofootprint group. Additionally, the biofootprint calculation is no longer used. Therefore we've chosen to simply remove the group for now, while awaiting the cleanup op the entire method in #3296