Restructure primary CO2 emission queries
Energetic use is applied to industry emissions, this was a bug that is fixed. Non energetic emissions in industry of nodes with group co2_emissions_refinery_products is added to non energetic query, including a correction for the free_co2_factor. This is needed since the weighted_carrier_co2_per_mj method does not take this into account.